requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
et_xmlfile==2.0.0
idna==3.10
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
scipy==1.15.2
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
